requirements.txt
fastapi
uvicorn
requests
ddgs
python-multipart
PyMuPDF
python-docx